Fix warnings, code clean up in CsNetwork

+++ src/network/access/qspdyprotocolhandler.cpp
@@ -1218,16 +1218,16 @@ void QSpdyProtocolHandler::handleDataFrame(const QByteArray &frameHeaders)
          && dataLeftInWindow < replyPrivate->windowSizeDownload / 2) {
 
       // socket read buffer size is 64K actually, hard coded in the channel
-      // We can read way more than 64K per socket, because the window size
-      // here is per stream.
+      // We can read way more than 64K per socket, because the window size here is per stream.
       if (replyPrivate->windowSizeDownload >= m_socket->readBufferSize()) {
          replyPrivate->windowSizeDownload = m_socket->readBufferSize();
       } else {
          replyPrivate->windowSizeDownload *= 1.5;
       }
+
       QMetaObject::invokeMethod(this, "sendWINDOW_UPDATE", Qt::QueuedConnection,
-                                Q_ARG(qint32, streamID),
-                                Q_ARG(quint32, replyPrivate->windowSizeDownload));
+                  Q_ARG(qint32, streamID), Q_ARG(quint32, static_cast<quint32>(replyPrivate->windowSizeDownload)));
+
       // setting the current data count to 0 is a race condition,
       // because we call sendWINDOW_UPDATE through the event loop.
       // But then again, the whole situation is a race condition because
